+++ backend/src/main/java/dev/be/dorothy/security/authentication/UsernameAndPasswordTokenProvider.java
@@ -1,7 +1,7 @@
-package dev.be.dorothy.auth.authentication;
+package dev.be.dorothy.security.authentication;
 
-import dev.be.dorothy.auth.MemberDetail;
-import dev.be.dorothy.auth.MemberDetailService;
+import dev.be.dorothy.security.context.MemberDetail;
+import dev.be.dorothy.security.context.MemberDetailService;
 import org.springframework.stereotype.Component;
 
 @Component
