+++ backend/src/main/java/dev/be/dorothy/security/authorization/ExpressionUrlAuthorizationConfigurer.java
@@ -1,4 +1,4 @@
-package dev.be.dorothy.auth.authorization;
+package dev.be.dorothy.security.authorization;
 import dev.be.dorothy.member.MemberRole;
 import org.springframework.stereotype.Component;
