+++ backend/src/main/java/dev/be/dorothy/security/authorization/SecurityConfig.java
@@ -1,6 +1,6 @@
-package dev.be.dorothy.auth.authorization;
+package dev.be.dorothy.security.authorization;
 
-import dev.be.dorothy.auth.AuthorizationFilter;
+import dev.be.dorothy.security.filter.AuthorizationFilter;
 import dev.be.dorothy.member.MemberRole;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
@@ -10,13 +10,13 @@ import org.springframework.core.annotation.Order;
 public class SecurityConfig {
 
     @Bean
-    @Order(2)
+    @Order(3)
     public AuthorizationFilter authorizationFilter(HttpSecurity httpSecurity){
         return httpSecurity
                 .authorizeRequest()
                 .antMatchers("/login").permitAll()
-                .antMatchers("/member").isAuthenticated()
-                .antMatchers("/track").hasRole(MemberRole.ADMIN)
+                .antMatchers("/track").isAuthenticated()
+                .antMatchers("/track/generate").hasRole(MemberRole.ADMIN)
                 .and().build();
     }
 }
