+++ backend/src/main/java/dev/be/dorothy/security/context/MemberDetail.java
@@ -1,4 +1,4 @@
-package dev.be.dorothy.auth;
+package dev.be.dorothy.security.context;
 
 import dev.be.dorothy.member.Member;
 import dev.be.dorothy.member.service.MemberResDto;
