+++ backend/src/test/java/dev/be/dorothy/security/SecurityFilterTest.java
@@ -1,6 +1,9 @@
-package dev.be.dorothy.auth;
+package dev.be.dorothy.security;
 
-import dev.be.dorothy.auth.authentication.UsernameAndPasswordTokenProvider;
+import dev.be.dorothy.security.authentication.AuthenticationManager;
+import dev.be.dorothy.security.authentication.UsernameAndPasswordTokenProvider;
+import dev.be.dorothy.security.filter.AuthenticationFilter;
+import dev.be.dorothy.security.filter.LoginFilter;
 import dev.be.dorothy.member.service.PasswordEncryptor;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
